fix compiler warnings

+++ util/array2d.h
@@ -38,8 +38,7 @@ enum LayoutCategory {
  *        or jagged arrays, in which the size along the second dimension
  *        is different for each element.
  */
-template <typename T, typename F = uint64_t>
-class array2d {
+template <typename T, typename F = uint64_t> class array2d {
  public:
   /**
    * \brief Constructs a 2d array, setting the layout and stride
@@ -100,7 +99,7 @@ class array2d {
    *
    * \return number of elements (such as number of vertices, or triangles).
    */
-  int64_t n() const {
+  size_t n() const {
     if (layout_ == Layout_Rectangular) {
       ASSERT(stride_ > 0);
       return data_.size() / stride_;
@@ -176,8 +175,7 @@ class array2d {
    *
    * \param[in] x - values to add (there must be stride elements)
    */
-  template <typename R>
-  void add(const R* x) {
+  template <typename R> void add(const R* x) {
     ASSERT(layout_ == Layout_Rectangular);
     for (int j = 0; j < stride_; j++) data_.push_back(x[j]);
   }
@@ -188,8 +186,7 @@ class array2d {
    * \param[in] x - values to add (there must be stride elements)
    * \param[in] n - number of values to add
    */
-  template <typename R>
-  void add(const R* x, int m) {
+  template <typename R> void add(const R* x, int m) {
     if (layout_ == Layout_Rectangular) {
       add(x);
       return;
